css: keep ns_css_prop_syntax returning a signed value
ns_css_prop_syntax mixes its int parameter with kProperty[].syntax, which
is the ns_css_prop enum and unsigned, so the conditional changed
signedness and -Wsign-compare rejected it. The Linux CI job builds with
-Werror and failed; local builds do not set -Werror, which is why it went
unnoticed.

The function returns int, so convert the enum arm explicitly.

+++ src/css.c
@@ -415,7 +415,7 @@ ns_css_prop_syntax(int prop)
 {
     if (prop < 0 || prop >= NS_CSS_PROP_COUNT) return prop;
     return kProperty[prop].syntax == NS_CSS_PROP_COUNT
-        ? prop : kProperty[prop].syntax;
+        ? prop : (int)kProperty[prop].syntax;
 }
 
 static int
